Added API & Location Select
Added API weerlive.nl  , free 300 request per day and 2434 dutch locations.
Added CSV for locations
Added Dropdown to select current location.

Fetch will first check if it has saved the weatherdata in the last 10min of the city, if not it will make the API call, otherwise it uses localstorage (this to limit the API calls) , unless location is switched non stop it should prevent running out of API calls.